@@ -1,4 +1,5 @@
-from typing import Iterable, Tuple, Union
+from collections.abc import Iterable
+from typing import Union
 
 from flair.data import Corpus, MultiCorpus
 from flair.models import MultitaskModel
@@ -6,18 +7,18 @@ from flair.nn import Classifier, Model
 
 
 def make_multitask_model_and_corpus(
-    mapping: Iterable[Union[Tuple[Classifier, Corpus], Tuple[Classifier, Corpus, float]]]
-) -> Tuple[Model, Corpus]:
+    mapping: Iterable[Union[tuple[Classifier, Corpus], tuple[Classifier, Corpus, float]]]
+) -> tuple[Model, Corpus]:
     models = []
     corpora = []
     loss_factors = []
     ids = []
 
-    for task_id, map in enumerate(mapping):
-        models.append(map[0])
-        corpora.append(map[1])
-        if len(map) == 3:
-            loss_factors.append(map[2])
+    for task_id, _map in enumerate(mapping):
+        models.append(_map[0])
+        corpora.append(_map[1])
+        if len(_map) == 3:
+            loss_factors.append(_map[2])
         else:
             loss_factors.append(1.0)
 
