@@ -1,19 +1,14 @@
 import logging
 from collections import defaultdict
+from collections.abc import Iterator, Sequence
 from inspect import isclass, signature
 from itertools import count
 from queue import Queue
 from typing import (
     Any,
     Callable,
-    Dict,
-    Iterator,
-    List,
     NewType,
     Optional,
-    Sequence,
-    Set,
-    Type,
     Union,
     cast,
 )
@@ -23,7 +18,7 @@ from flair.distributed_utils import is_main_process
 log = logging.getLogger("flair")
 
 
-PluginArgument = Union["BasePlugin", Type["BasePlugin"]]
+PluginArgument = Union["BasePlugin", type["BasePlugin"]]
 HookHandleId = NewType("HookHandleId", int)
 
 EventIdenifier = str
@@ -36,7 +31,7 @@ class TrainingInterrupt(Exception):
 class Pluggable:
     """Dispatches events which attached plugins can react to."""
 
-    valid_events: Optional[Set[EventIdenifier]] = None
+    valid_events: Optional[set[EventIdenifier]] = None
 
     def __init__(self, *, plugins: Sequence[PluginArgument] = []) -> None:
         """Initialize a `Pluggable`.
@@ -44,11 +39,11 @@ class Pluggable:
         Args:
             plugins: Plugins which should be attached to this `Pluggable`.
         """
-        self._hook_handles: Dict[EventIdenifier, Dict[HookHandleId, HookHandle]] = defaultdict(dict)
+        self._hook_handles: dict[EventIdenifier, dict[HookHandleId, HookHandle]] = defaultdict(dict)
 
         self._hook_handle_id_counter = count()
 
-        self._plugins: List[BasePlugin] = []
+        self._plugins: list[BasePlugin] = []
 
         # This flag tracks, whether an event is currently being processed (otherwise it is added to the queue)
         self._processing_events = False
@@ -183,7 +178,7 @@ class BasePlugin:
 
     def __init__(self) -> None:
         """Initialize the base plugin."""
-        self._hook_handles: List[HookHandle] = []
+        self._hook_handles: list[HookHandle] = []
         self._pluggable: Optional[Pluggable] = None
 
     def attach_to(self, pluggable: Pluggable):
@@ -269,7 +264,7 @@ class BasePlugin:
     def __str__(self) -> str:
         return self.__class__.__name__
 
-    def get_state(self) -> Dict[str, Any]:
+    def get_state(self) -> dict[str, Any]:
         return {"__cls__": f"{self.__module__}.{self.__class__.__name__}"}
 
 
