@@ -1,6 +1,6 @@
 import logging
 from pathlib import Path
-from typing import Any, Dict
+from typing import Any
 
 from flair.trainers.plugins.base import TrainerPlugin
 from flair.training_utils import add_file_handler
@@ -25,5 +25,5 @@ class LogFilePlugin(TrainerPlugin):
     def attach_to_all_processes(self) -> bool:
         return False
 
-    def get_state(self) -> Dict[str, Any]:
+    def get_state(self) -> dict[str, Any]:
         return {**super().get_state(), "base_path": str(self.base_path)}
