@@ -5,7 +5,7 @@ from enum import Enum
 from functools import reduce
 from math import inf
 from pathlib import Path
-from typing import Dict, List, Literal, Optional, Union
+from typing import Literal, Optional, Union
 
 from scipy.stats import pearsonr, spearmanr
 from sklearn.metrics import mean_absolute_error, mean_squared_error
@@ -25,7 +25,7 @@ class Result:
         main_score: float,
         detailed_results: str,
         classification_report: Optional[dict] = None,
-        scores: Optional[Dict] = None,
+        scores: Optional[dict] = None,
     ) -> None:
         classification_report = classification_report if classification_report is not None else {}
         assert scores is not None and "loss" in scores, "No loss provided."
@@ -47,8 +47,8 @@ class MetricRegression:
     def __init__(self, name) -> None:
         self.name = name
 
-        self.true: List[float] = []
-        self.pred: List[float] = []
+        self.true: list[float] = []
+        self.pred: list[float] = []
 
     def mean_squared_error(self):
         return mean_squared_error(self.true, self.pred)
@@ -98,7 +98,7 @@ class WeightExtractor:
         if isinstance(directory, str):
             directory = Path(directory)
         self.weights_file = init_output_file(directory, "weights.txt")
-        self.weights_dict: Dict[str, Dict[int, List[float]]] = defaultdict(lambda: defaultdict(list))
+        self.weights_dict: dict[str, dict[int, list[float]]] = defaultdict(lambda: defaultdict(list))
         self.number_of_weights = number_of_weights
 
     def extract_weights(self, state_dict, iteration):
@@ -338,7 +338,7 @@ def init_output_file(base_path: Union[str, Path], file_name: str) -> Path:
     return file
 
 
-def convert_labels_to_one_hot(label_list: List[List[str]], label_dict: Dictionary) -> List[List[int]]:
+def convert_labels_to_one_hot(label_list: list[list[str]], label_dict: Dictionary) -> list[list[int]]:
     """Convert list of labels to a one hot list.
 
     Args:
@@ -365,9 +365,9 @@ def add_file_handler(log, output_file):
 
 
 def store_embeddings(
-    data_points: Union[List[DT], Dataset],
+    data_points: Union[list[DT], Dataset],
     storage_mode: EmbeddingStorageMode,
-    dynamic_embeddings: Optional[List[str]] = None,
+    dynamic_embeddings: Optional[list[str]] = None,
 ):
     if isinstance(data_points, Dataset):
         data_points = list(_iter_dataset(data_points))
@@ -391,7 +391,7 @@ def store_embeddings(
             data_point.to("cpu", pin_memory=pin_memory)
 
 
-def identify_dynamic_embeddings(data_points: List[DT]) -> Optional[List]:
+def identify_dynamic_embeddings(data_points: list[DT]) -> Optional[list[str]]:
     dynamic_embeddings = []
     all_embeddings = []
     for data_point in data_points:
